Create separate module containing common test code
