+++ app/src/test/java/com/alvindizon/tampisaw/domain/GetAllPhotosUseCaseTest.kt
@@ -1,7 +1,7 @@
 package com.alvindizon.tampisaw.domain
 
 import com.alvindizon.tampisaw.TestConstants
-import com.alvindizon.tampisaw.collectData
+import com.alvindizon.tampisaw.testbase.collectData
 import io.mockk.every
 import io.mockk.mockk
 import io.reactivex.rxjava3.core.Observable
