+++ build.gradle
@@ -25,6 +25,17 @@ plugins {
 
 allprojects {
     addRepos(repositories)
+    tasks.withType(org.jetbrains.kotlin.gradle.tasks.KotlinCompile).all {
+        kotlinOptions {
+            allWarningsAsErrors = true
+            jvmTarget = "1.8" // this needs to be set if you want your decompiled Java code at Java 8
+            freeCompilerArgs += [
+                    "-Xjvm-default=all",
+                    "-Xopt-in=kotlinx.coroutines.ExperimentalCoroutinesApi",
+                    "-Xopt-in=androidx.paging.ExperimentalPagingApi"
+            ]
+        }
+    }
 }
 
 subprojects {
