+++ common/testbase/src/main/java/com/alvindizon/tampisaw/testbase/TestUtils.kt
@@ -1,4 +1,4 @@
-package com.alvindizon.tampisaw
+package com.alvindizon.tampisaw.testbase
 
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.Observer
